requirements.txt
aiohttp
asyncio
dashscope
google-generativeai
llama-index
nest-asyncio
openai
qianfan
pyyaml
tenacity
tiktoken
urllib3